nochangelog: Extract config parsing into pkg/config
